Put Saldo done (tested) Pull Saldo and Distance done (not tested)

+++ DriverService/Models/PriceAdmin.cs
@@ -8,7 +8,7 @@ namespace DriverService.Models
     public partial class PriceAdmin
     {
         public int PriceId { get; set; }
-        public double Price { get; set; }
+        public float Price { get; set; }
         public DateTime Created { get; set; }
     }
 }
